+++ src/components/DiagnosticsServiceTabs.tsx
@@ -43,6 +43,7 @@ export default class DiagnosticsServiceTabs extends TabContainer<DiagnosticsServ
     // If the collection has any timestamps with exceptions, it should start out expanded.
     return Object.keys(collections).map((collectionName) =>
       <Panel
+        key={collectionName}
         headerText={collectionName}
         openByDefault={collections[collectionName].some((ts) => ts.exception)}
         content={this.renderTimestamps(collections[collectionName])}