+++ src/components/TroubleshootingPage.tsx
@@ -1,0 +1,74 @@
+import * as React from "react";
+import { Store } from "redux";
+import * as PropTypes from "prop-types";
+import Header from "./Header";
+import { State } from "../reducers/index";
+import TroubleshootingTabContainer from "./TroubleshootingTabContainer";
+
+
+export interface TroubleshootingPageContext {
+  editorStore: Store<State>;
+  csrfToken: string;
+}
+
+export interface TroubleshootingPageState {
+  tab: string;
+  subtab: string;
+}
+
+export interface TroubleshootingPageProps extends React.Props<TroubleshootingPageProps> {
+  params: {
+    tab?: string;
+    subtab?: string;
+  };
+}
+
+export default class TroubleshootingPage extends React.Component<TroubleshootingPageProps, TroubleshootingPageState> {
+  context: TroubleshootingPageContext;
+
+  static contextTypes: React.ValidationMap<TroubleshootingPageContext> = {
+    editorStore: PropTypes.object.isRequired,
+    csrfToken: PropTypes.string.isRequired
+  };
+
+  CATEGORIES = {
+    "diagnostics": ["coverage_provider", "script", "monitor", "other"],
+    "self-tests": ["collections", "patronAuthServices", "searchServices"]
+  };
+
+  constructor(props) {
+    super(props);
+    this.state = { tab: "diagnostics", subtab: "coverage_provider" };
+    this.goToTab = this.goToTab.bind(this);
+  }
+
+  componentWillMount() {
+    document.title = "Circulation Manager - Troubleshooting";
+  }
+
+  render(): JSX.Element {
+    let tab = this.props.params.tab || this.state.tab;
+    let subtab = this.props.params.subtab || this.state.subtab;
+    subtab = this.CATEGORIES[tab].indexOf(subtab) >= 0 ? subtab : this.CATEGORIES[tab][0];
+    return(
+      <div className="troubleshooting-page">
+        <Header />
+        <div className="body">
+          <h2>Troubleshooting</h2>
+          <TroubleshootingTabContainer
+            store={this.context.editorStore}
+            csrfToken={this.context.csrfToken}
+            tab={tab}
+            goToTab={this.goToTab}
+            subtab={subtab}
+          />
+        </div>
+      </div>
+    );
+  }
+
+  goToTab(tab: string) {
+    this.setState({ tab, subtab: this.props.params.subtab });
+  }
+
+}