+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 ## Changelog
 
 ### v0.3.11
+#### Added
+- Created a Troubleshooting page, comprising Diagnostics and Self-tests.
+#### Updated
+- Made the buttons, icons, and links on the catalog page visually consistent with the rest of the interface.
 #### Fixed
 - Fixed an issue with the alignment of delete buttons in the Input List component.